refactor: Enabling heuristic metrics from Opik, updating documentation, and minor fixes to uncertainty metrics
